Split Generation algorithm into upsert_resolution.rs
Move the upsert-resolution algorithm (Generation, UpsertE/EV,
FinalPopulations, plus their helpers) into a new src/upsert_resolution.rs,
mirroring Mentat's split. The driver — `resolve_tempids_with_upserts` —
stays in tx.rs as a thin orchestrator over the Generation API, alongside
the partition-decision helpers (`infer_tempid_partitions`,
`allocation_partitions`) and the `resolve_temp_id_avs` /
`record_resolutions` callbacks the loop drives.

Generation-related types are exposed `pub(crate)` for the driver, and the
unused `is_identity_attr` is dropped.
